+++ clients/admin-ui/src/features/data-discovery-and-detection/tables/DetectionResultTable.tsx
@@ -87,7 +87,12 @@ const DetectionResultTable = ({ resourceUrn }: MonitorResultTableProps) => {
     endRange,
     pageIndex,
     setTotalPages,
+    resetPageIndexToDefault,
   } = useServerSidePagination();
+
+  useEffect(() => {
+    resetPageIndexToDefault();
+  }, [resourceUrn, searchQuery, resetPageIndexToDefault]);
 
   const {
     isFetching,